requirements.txt
nltk
numpy
pandas
pprintpp
tqdm
bs4
matplotlib
requests
ratelimit
scikit-learn
editdistance
yfinance
statsmodels
pickleshare
pandas_datareader
lxml
openai
scipy
flask
flask-cors
python-dotenv
chromadb